Disable DNS-cache for libcurl
When SPID moved to Amazon and started to use Amazon Elastic Loadbalancer it broke a lot of our serverside scripts. libcurl has it's own builtin DNS-cache for each thread that is set to 120 Seconds. It addition it has a bug that are present in quite a lot of the older libcurl-installations that never times out the dns-cache if the thread is very busy